--- Test/test_multiroom_rooms/index.html
@@ -1,95 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-<title>Void</title>
-<meta charset="utf-8">
-<meta name="viewport" content="width=device-width, initial-scale=1">
-<link rel="stylesheet" href="http://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
-
-<style>
-body {
-  background-color: #1a1a1a;
-  font-family: ‘Lucida Console’, Monaco, monospace;
-}
-.background {
-  background-color: #1a1a1a;
-}
-form {
-  padding-top: 35vh;
-  /*height: 50%;*/
-  color: #ffffff;
-  background-color: #1a1a1a;
-  text-align: center;
-  vertical-align: middle;
-
-}
-h1 {
-  padding-top: 5vh;
-  padding-left: 5vw;
-  font-family: ‘Lucida Console’, Monaco, monospace;
-}
-/*p {
-
-}*/
-</style>
-</head>
-
- <body>
-
-   <div class="background">
-     <h1>Void</h1>
-     <div class="input-section">
-         <!--
-         action="chatroom.html"
-         -->
-       <form>
-         <!--
-           <input id="room" type="text" autocomplete="off" style="background-color: #1a1a1a; text-align: center;" name="RoomID" placeholder="Room ID"><br><br>
-         -->
-         <input id="alias" type="text" autocomplete="off" style="background-color: #1a1a1a; text-align: center;" name="alias" placeholder="Alias"><br><br>
-         <input formaction ="room1.html" onclick="joinRoom1()" type="submit" autocomplete="off" style="background-color: #1a1a1a; stroke-width: 0;" value="Room1">
-         <input formaction ="room2.html" onclick="joinRoom2()" type="submit" autocomplete="off" style="background-color: #1a1a1a; stroke-width: 0;" value="Room2">
-         <input formaction ="room3.html" onclick="joinRoom3()" type="submit" autocomplete="off" style="background-color: #1a1a1a; stroke-width: 0;" value="Room3">
-         <!--<input type="submit" autocomplete="off" style="background-color: #1a1a1a; stroke-width: 0;" value="Enter Room">
-            -->
-       </form>
-     </div>
-   </div>
- </body>
-<script src="/socket.io/socket.io.js"></script>
-<script src="http://code.jquery.com/jquery-1.11.1.js"></script>
-<script>
-  var socket = io();
-  var alias;
-  alias = $('#alias').val();
-
-  function joinRoom1(){
-      socket.emit('joinRoom',alias,"room1");
-  }
-  function joinRoom2(){
-      socket.emit('joinRoom',alias,"room2");
-  }
-  function joinRoom3(){
-      socket.emit('joinRoom',alias,"room3");
-  }
-  $('form').submit(function(){
-      //alias = $('#alias').val();
-
-      /*
-    room = $('#room').val();
-    alias = $('#alias').val();
-      /*
-       Should add checks here if the room and alias entered are blank
-
-      socket.emit('joinRoom',room);
-      //console.log("Called 'joinRoom'");
-      //socket.emit('createalias',alias);
-      //socket.emit('joinRoom',room);
-      //socket.emit('enterRoom');
-
-      //socket.emit('joinRoom');
-       */
-  });
-</script>
- </body>
-</html>
