--- Test/test_oneroom/index.html
@@ -1,65 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <title>Void</title>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="http://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
-
-    <style>
-        body {
-            background-color: #1a1a1a;
-            font-family: ‘Lucida Console’, Monaco, monospace;
-        }
-        .background {
-            background-color: #1a1a1a;
-        }
-        form {
-            padding-top: 35vh;
-            /*height: 50%;*/
-            color: #ffffff;
-            background-color: #1a1a1a;
-            text-align: center;
-            vertical-align: middle;
-        }
-        h1 {
-            padding-top: 5vh;
-            padding-left: 5vw;
-            font-family: ‘Lucida Console’, Monaco, monospace;
-        }
-        /*p {
-        }*/
-    </style>
-</head>
-
-<body>
-
-<div class="background">
-    <h1>Void</h1>
-    <div class="input-section">
-        <form action="chatroom.html">
-            <!--<input id="room" type="text" autocomplete="off" style="background-color: #1a1a1a; text-align: center;" name="RoomID" placeholder="Room ID"><br><br>
-            -->
-            <input id="alias" type="text" autocomplete="off" style="background-color: #1a1a1a; text-align: center;" name="alias" placeholder="Alias"><br><br>
-            <input type="submit" autocomplete="off" style="background-color: #1a1a1a; stroke-width: 0;" value="Enter Chat">
-        </form>
-    </div>
-
-</div>
-</body>
-<script src="/socket.io/socket.io.js"></script>
-<script src="http://code.jquery.com/jquery-1.11.1.js"></script>
-<script>
-    var socket = io.connect('http://localhost:5000');
-    jQuery(function($){
-        var $alias = $('#alias');
-    });
-
-
-    $('form').submit(function(){
-        socket.emit('newuser',alias);
-    });
-</script>
-
-</body>
-</html>
