--- Test/test_staticroom_namespaces/chat.css
@@ -1,10 +1,0 @@
-.chat-log{
-    float:left;
-    height: 90vh;
-    width: 98vw;
-    margin-top: 1vh;
-    margin-left: 1vw;
-    margin-right: 1vw;
-    background-color: #1a1a1a;
-    color: #f2f2f2;
-}
