--- Test/test_staticroom_namespaces/room3.html
@@ -1,79 +1,0 @@
-<!doctype html>
-<html>
-<head>
-    <title> VOID_Room3 </title>
-    <style>
-        * {color: white; background: #1a1a1a; margin: 0; padding: 0; box-sizing: border-box ; }
-        body {font: 13px Lucida Console, monospace;}
-        form {background: #1a1a1a ; padding: 3px; position: fixed; bottom: 0; width: 100%; }
-    </style>
-</head>
-<body>
-<div id = "chatlog"></div>
-<h1 id="h1">Room3</h1>
-<form id="form" action="" >
-    <input id="msg" autocomplete="off" type="text"
-           style="height: 3vh;
-           color: #f2f2f2;
-           width: 87vw;
-           background-color: #1a1a1a;
-           display: inline-block;
-           float: left;
-           padding-left: 1vw;
-           padding-right: 1vw;"
-           name="message-input"
-           value=""
-    >
-    <input id="sendButton" type="button" autocomplete="off" onclick="submitMessage()" name="SendButton" value = "Send"
-           style = "width: 4vw;
-           color: #f2f2f2;
-           background-color: #1a1a1a;
-           display: inline-block;
-           float: left;
-           text-align: center;
-           height: 3vh;"
-    >
-</form>
-<script src="/socket.io/socket.io.js"></script>
-<script src="http://code.jquery.com/jquery-1.11.1.js"></script>
-<script>
-
-    var dir = "/room3";
-    var socket = io(dir);
-
-    /*
-     LISTENER
-     */
-    socket.on('updateChat',function(data,user){
-        console.log("updateChat called: message = "+ data);
-        var newline = "<br>";
-        var message = data;
-        console.log("Attempting to update Chatlog");
-        document.body.innerHTML = document.body.innerHTML + user + ": " + message + newline;
-    });
-    socket.on('updateRoomName',function(roomname){
-        var room = roomname;
-        document.getElementById("h1").innerHTML = room;
-
-    });
-    /*
-     Submit Function
-     */
-    function submitMessage(){
-        /* Local Update of HTML Div - Used this for testing
-         var newline = "<br>";
-         var message = $('#msg').val();
-         document.body.innerHTML = document.body.innerHTML + message + newline;
-         */
-        var message = $('#msg').val();
-        socket.emit("sendmsg",message);
-        console.log("Called 'sendmsg'");
-
-    }
-</script>
-</body>
-</html>
-
-<!--
-#1a1a1a
--->
